Joy Ride
This is a quick, simple project that gives you a chance to write code that controls a simulated car as you get familiar with the Workspaces you'll be using throughout this Nanodegree program. This project has three parts (but you will only submit part three).

Part One - Drag Race: In this part, you'll write code that lets a car jump over a grove of trees. This might not be a common scenario for a self-driving car, but it will get you familiar with the programming interface.
Part Two - Circular Track: In this part, you'll write code that lets a car navigate a circular track. In doing so you'll explore the relationship between steering angle and turning radius.
Part Three - Parallel Park: In this part, you'll write a sequence of instructions that successfully parallel parks a car. In a real car, you obviously wouldn't be able to try repeatedly and hit other cars while you perfect your parking, but in this project, you can try as many times as you need to!